added neighborhood ops
